# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot follow yourself or already following
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{userId}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get user followers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: List of followers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users that a user is following
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: List of users being followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/me/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload avatar image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required:
                - avatar
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        bio:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        posts_count:
          type: integer
        followers_count:
          type: integer
        following_count:
          type: integer
        is_following:
          type: boolean
          description: Whether the current user follows this user
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - email
        - first_name
        - last_name
        - posts_count
        - followers_count
        - following_count
        - is_following
        - created_at
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_next:
          type: boolean
      required:
        - users
        - total
        - page
        - limit
        - has_next
    UpdateProfileRequest:
      type: object
      properties:
        bio:
          type: string
          maxLength: 160
          nullable: true
        avatar_url:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        first_name:
          type: string
        last_name:
          type: string
    AvatarUploadResponse:
      type: object
      properties:
        avatar_url:
          type: string
        message:
          type: string
      required:
        - avatar_url
        - message
...