# lib/api-zod/src/generated/types/avatarUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AvatarUploadResponse {
  avatar_url: string;
  message: string;
}