# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  /** @maxLength 160 */
  bio?: string | null;
  avatar_url?: string | null;
  website?: string | null;
  location?: string | null;
  first_name?: string;
  last_name?: string;
}